+++ src/pages/Properties/components/GCHKWarehouseMLotCodePrintProperties.jsx
@@ -1,5 +1,6 @@
 import EntityScanProperties from "./entityProperties/EntityScanProperties";
 import GCHKWarehouseMLotCodePrintTable from "../../../components/Table/gc/GCHKWarehouseMLotCodePrintTable";
+import TableManagerRequest from "../../../api/table-manager/TableManagerRequest";
 
 
 export default class GCHKWarehouseMLotCodePrintProperties extends EntityScanProperties{
@@ -17,6 +18,22 @@ export default class GCHKWarehouseMLotCodePrintProperties extends EntityScanProp
         loading: false,
         resetFlag: true
       });
+  }
+
+  queryData = (whereClause) => {
+    const self = this;
+    let requestObject = {
+      tableRrn: this.state.tableRrn,
+      whereClause: whereClause,
+      success: function(responseBody) {
+        self.setState({
+          tableData: responseBody.dataList,
+          loading: false
+        });
+        self.form.resetFormFileds();  
+      }
+    }
+    TableManagerRequest.sendGetDataByRrnRequest(requestObject);
   }
 
     buildTable = () => {
